+++ example/storybook/src/api/DescendantsStyles/ContextBasedStyles.tsx
@@ -1,20 +1,86 @@
 import React from 'react';
 
-import { Pressable, Text, View } from 'react-native';
+import { Pressable, Text, TextInput } from 'react-native';
 import { AsForwarder, styled } from '@gluestack-style/react';
 import { Wrapper } from '../../components/Wrapper';
-import { AddIcon } from '@gluestack/design-system';
+// import { AddIcon } from '@gluestack/design-system';
+import { createIcon } from '@gluestack-ui/icon';
+import { Svg } from 'react-native-svg';
 
-const Box = styled(View, {
-  bg: '$red500',
-  width: '100px',
-  height: '100px',
-});
+const StyledButton = styled(
+  Pressable,
+  {
+    backgroundColor: '$primary500',
+    p: '$2',
+    // _icon: {
+    //   props: {
+    //     color: '$blue500',
+    //   },
+    // },
+  },
+  {
+    descendantStyle: ['_icon'],
+  }
+);
+
+const StyledIcon = styled(
+  Text,
+  {
+    color: 'red',
+    props: {
+      placeholderTextColor: 'red',
+    },
+    _dark: {
+      props: {
+        placeholderTextColor: 'red',
+      },
+    },
+  },
+  {
+    // ancestorStyle: ['_icon'],
+    DEBUG: 'STYLED_ICON',
+  }
+);
+
+const MyText = styled(
+  StyledIcon,
+  {
+    color: 'blue',
+    bg: '$amber400',
+    // props: {
+    //   color: '$white',
+    // },
+  },
+  {
+    // ancestorStyle: ['_icon'],
+    DEBUG: 'MYTEXT',
+  }
+);
+
+const MyTextForward = styled(
+  AsForwarder,
+  {
+    color: 'yellow',
+    // props: {
+    //   color: '$white',
+    // },
+  },
+  {
+    // ancestorStyle: ['_icon'],
+    DEBUG: 'MYTEXT',
+  }
+);
 
 export function ContextBasedStyles() {
   return (
     <Wrapper>
-      <Box sx={{ bg: '$amber500' }} />
+      <StyledButton>
+        <MyText>hello</MyText>
+        {/* <StyledIcon as={MyText}>Text</StyledIcon>
+        <MyText as={StyledIcon}>Text</MyText> */}
+        {/* <MyTextForward as={StyledIcon}>Text</MyTextForward> */}
+        {/* <MyText>Text</MyText> */}
+      </StyledButton>
     </Wrapper>
   );
 }