+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gluestack-style/react
 
+## 0.1.32
+
+### Patch Changes
+
+- - Prop resolution issue when composing on styled component into another.
+
+## 0.1.31
+
+### Patch Changes
+
+- - Global style media query typing fixes
+
 ## 0.1.30
 
 ### Patch Changes